+++ src/resources/lang/en/messages.php
@@ -150,4 +150,7 @@ return array(
         => "That field is not supported. Please try again using the dropdown list.",
     'no_revision_found'
         => "No revision history.",
+    'product_variant_inherits_from_parent_footnote'
+        => "Variant inherits the same category as the main product. "
+            . "If you don't see the category here, go back to the main product and save the changes first.",
 );